component refactored
