+++ client/src/components/Account.jsx
@@ -1,45 +1,35 @@
-import React from 'react';
 import axios from 'axios';
 import Playlists from './Playlists.jsx';
 import PlayPlaylist from './PlayPlaylist.jsx';
+import React, {useState, useEffect} from 'react';
 
-class Account extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      profile: {},
-      playlists: [],
-      tracks: [],
-      currentPlaying: {}
-    };
-    this.userProfile = this.userProfile.bind(this);
-    this.getPlaylists = this.getPlaylists.bind(this);
-    this.onePlaylist = this.onePlaylist.bind(this);
-    this.playPlaylist = this.playPlaylist.bind(this);
-  }
-  userProfile() {
+
+const AccountHooks = (props) => {
+  const [playlists, setPlaylists] = useState([]);
+  const [profile, setProfile] = useState({});
+  const [currentlyPlaying, setCurrentlyPlaying] = useState({});
+ 
+  const userProfile = () => {
     axios({
       method: 'GET',
       url: '/me',
     }).then(({data}) => {
-      this.setState({
-        profile: data
-      });
+      setProfile(data);
     }).catch((err) => console.error('err'));
-  }
+  };
 
-  getPlaylists () {
+  const getPlaylists = () =>{
+    console.log('getPlaylists');
     axios({
       method: 'GET',
       url: '/playlist',
     }).then(({data}) => {
-      this.setState({
-        playlists: data
-      });
+      setPlaylists(data);
+      console.log('playlist', playlists);
       return data;
     }).then((data) => {
       // things I want to share
-      const user = this.state.profile.display_name;
+      const user = profile.display_name;
       const playlistInfo = data.map((playlist) => {
         return {
           playlistIDs: playlist.id,
@@ -62,27 +52,12 @@ class Account extends React.Component {
         data: data
       });
     });
-  }
+  };
 
-  // not going to be used
-  onePlaylist(playlist) {
-    // console.log('click', playlist.id);
-    axios({
-      method: 'GET',
-      url: `/playlist/${playlist.id}`,
-    }).then(({data}) => {
-      // console.log(data);
-      this.setState({
-        tracks: data.items,
-      });
-    });
-  }
 
-  playPlaylist(playlist) {
-    // console.log(playlist);
-    this.setState({
-      playPlaylist: playlist
-    });
+  const playPlaylist = (playlist) =>{
+    console.log(playlist);
+    setCurrentlyPlaying(playlist);
     axios({
       method: 'GET',
       url: `/playlist/${playlist.id}`,
@@ -110,23 +85,24 @@ class Account extends React.Component {
         });
       }
     });
-  }
+  };
 
-  componentDidMount() {
-    this.userProfile();
-    this.getPlaylists();
-  }
+  // getPlaylists();
+  // userProfile();
 
-  render () {
-    const {profile, playlists, current, playPlaylist} = this.state;
-    return (
-      <div>
-        <h4 style={{fontStyle: 'italic'}}>User: {profile.display_name}</h4>
-        <PlayPlaylist playlist={playPlaylist}/>
-        <Playlists handleClick={this.playPlaylist} playlists={playlists} />
-      </div>
-    );
-  }
-}
+  useEffect(() => { getPlaylists(), userProfile(); }, []);
 
-export default Account;
+
+
+  return (
+    <div>
+      <h4 style={{fontStyle: 'italic'}}>User: {profile.display_name}</h4>
+      {console.log(currentlyPlaying)}
+      {/* <PlayPlaylist playlist={currentlyPlaying}/> */}
+      {currentlyPlaying.id ? <PlayPlaylist currentlyPlaying={currentlyPlaying}/> : null}
+      <Playlists handleClick={playPlaylist} playlists={playlists} />
+    </div>
+  );
+};
+
+export default AccountHooks;
